requirements.txt
cryptography
Django
gunicorn
django-allauth
pandas==2.3.1
numpy==2.1.2
pillow==11.0.0
opencv-python==4.12.0.88
tqdm==4.67.1
jinja2==3.1.4
pyyaml==6.0.2
fsspec==2024.6.1
networkx==3.3
sympy==1.13.1
typing-extensions==4.12.2
dotenv
djangorestframework
djangorestframework-simplejwt
drf-yasg
psycopg2-binary
requests
uvicorn
whitenoise


# openai
langchain-openai
langchain 
langchain-community 
langchain-core 
langchainhub
langchain-ollama


#
